requirements.txt
aiml==0.9.2
aiohttp==3.8.5
argcomplete==3.6.3
beautifulsoup4==4.14.3
botorch==0.16.1
checklist==0.0.11
clearml==2.1.2
coloredlogs==15.0.1
contractions==0.1.73
ctc_score==0.1.3
datasets==4.4.2
docformatter==1.7.7
dppy==0.3.3
editdistance==0.8.1
emoji==2.15.0
fairscale==0.4.13
fairseq==0.12.2
faiss==1.5.3
filelock==3.20.3
fire==0.7.1
flair==0.15.1
fuzzywuzzy==0.18.0
fvcore==0.1.5.post20221221
GitPython==3.1.46
google_api_python_client==2.187.0
gpytorch==1.14.3
hydra-core==1.3.2
iopath==0.1.10
ipdb==0.13.13
ipython==8.12.3
joblib==1.5.3
jsonlines==4.0.0
langdetect==1.0.9
language_tool_python==3.2.2
litellm==1.80.15
lru==0.1
Markdown==3.10
matplotlib==3.10.8
mephisto==1.1.1
more_itertools==10.8.0
nltk==3.9.2
numpy==2.4.1
omegaconf==2.3.0
openai==2.15.0
packaging==25.0
pandas==2.3.3
Pillow==12.1.0
pinyin==0.4.0
potime==0.0.3
protobuf==3.20.3
pytest==9.0.2
pytest_regressions==2.9.1
pytorch_pretrained_bert==0.6.2
PyYAML==6.0.3
regex==2025.11.3
Requests==2.32.5
rouge==1.0.1
rouge_score==0.1.2
sacrebleu==2.5.1
scikit_learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
sentence_transformers==5.2.0
setuptools==68.0.0
sh==2.2.2
simcse==0.4
spacy==3.8.11
sphinx_rtd_theme==3.0.2
stanza==1.11.0
subword_nmt==0.3.8
tensorboardX==2.6
tensorboardX==2.6.4
tensorflow==2.20.0
torch_xla==2.9.0
torchtext==0.18.0
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.36.2
typing_extensions==4.15.0
Unidecode==1.4.0
vllm==0.13.0
wandb==0.23.1
wget==3.2
xmltodict==1.0.2
zstandard==0.25.0
